# backend/openapi.yaml
openapi: 3.1.0
info:
    title: VidyaSetu
    version: 0.1.0
    description: |
        VidyaSetu is a role-based education and workshop management backend API designed for Admin, Institutional Admin, Educator, Student, and Public users.

        The system supports the complete academic lifecycle including authentication, content delivery, assessments, certification, analytics, communication, and AI-powered features.

        #### Core capabilities include:

        - **Secure JWT-based authentication** and role-based access control for admin, institutional admin, educator, and student users.
        - **Institution and workshop administration**, including workshop creation, updates, enrollment tracking, assignments, and lifecycle management.
        - **Centralized content management** for modules, study materials, and learning resources uploaded by educators and accessed by students.
        - **Assessment management** for quizzes, tests, and assignments, including question creation, student attempts, answer submission, and result storage.
        - **Automated evaluation support** for objective assessments to reduce manual grading effort and improve consistency.
        - **Certificate generation** for successful workshop completion and public certificate verification using unique verification codes or identifiers.
        - **Performance tracking and reporting** for student progress, workshop statistics, assessment outcomes, completion rates, and trend analysis.
...

...

...

...

...

...

...

...

...
components:
    schemas:
...
        GradeResult:
            description: Returned by POST /tests/{id}/submit.
            properties:
                pass_fail:
                    title: Pass Fail
                    type: boolean
                per_question:
                    items:
                        additionalProperties: true
                        type: object
                    title: Per Question
                    type: array
                percentage:
                    title: Percentage
                    type: number
                score:
                    title: Score
                    type: integer
                submission_id:
                    title: Submission Id
                    type: string
                total_marks:
                    title: Total Marks
                    type: integer
            required:
                - submission_id
                - score
                - total_marks
                - percentage
                - pass_fail
                - per_question
            title: GradeResult
            type: object
        HTTPValidationError:
            properties:
                detail:
                    items:
                        $ref: "#/components/schemas/ValidationError"
                    title: Detail
                    type: array
            title: HTTPValidationError
            type: object
...
        NotificationBulkCreate:
            properties:
                message:
                    title: Message
                    type: string
                notification_type:
                    $ref: "#/components/schemas/NotificationType"
                    default: general
                user_ids:
                    items:
                        type: string
                    title: User Ids
                    type: array
            required:
                - user_ids
                - message
            title: NotificationBulkCreate
            type: object
        NotificationResponse:
            properties:
                created_at:
                    anyOf:
                        - format: date-time
                          type: string
                        - type: "null"
                    title: Created At
                id:
                    title: Id
                    type: string
                message:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Message
                notification_type:
                    anyOf:
                        - $ref: "#/components/schemas/NotificationType"
                        - type: "null"
                status:
                    anyOf:
                        - $ref: "#/components/schemas/NotificationStatus"
                        - type: "null"
                user_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: User Id
            required:
                - id
                - user_id
                - message
                - status
                - notification_type
                - created_at
            title: NotificationResponse
            type: object
        NotificationStatus:
            enum:
                - unread
                - read
            title: NotificationStatus
            type: string
        NotificationType:
            enum:
                - general
                - test
                - fees
                - attendance
                - certificate
            title: NotificationType
            type: string
        OptionItem:
            properties:
                id:
                    title: Id
                    type: string
                is_correct:
                    default: false
                    title: Is Correct
                    type: boolean
                text:
                    title: Text
                    type: string
            required:
                - id
                - text
            title: OptionItem
            type: object
        OptionItemPublic:
            properties:
                id:
                    title: Id
                    type: string
                text:
                    title: Text
                    type: string
            required:
                - id
                - text
            title: OptionItemPublic
            type: object
...
        Page_NotificationResponse_:
            properties:
                items:
                    items:
                        $ref: "#/components/schemas/NotificationResponse"
                    title: Items
                    type: array
                limit:
                    title: Limit
                    type: integer
                offset:
                    title: Offset
                    type: integer
                total:
                    title: Total
                    type: integer
            required:
                - items
                - total
                - offset
                - limit
            title: Page[NotificationResponse]
            type: object
...
        QuestionPublicResponse:
            properties:
                assessment_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Assessment Id
                id:
                    title: Id
                    type: string
                marks:
                    anyOf:
                        - type: integer
                        - type: "null"
                    title: Marks
                options:
                    default: []
                    items:
                        $ref: "#/components/schemas/OptionItemPublic"
                    title: Options
                    type: array
                text:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Text
                type:
                    anyOf:
                        - $ref: "#/components/schemas/QuestionType"
                        - type: "null"
            required:
                - id
                - assessment_id
                - text
                - type
                - marks
            title: QuestionPublicResponse
            type: object
        QuestionResponse:
            properties:
                assessment_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Assessment Id
                id:
                    title: Id
                    type: string
                marks:
                    anyOf:
                        - type: integer
                        - type: "null"
                    title: Marks
                options:
                    default: []
                    items:
                        $ref: "#/components/schemas/OptionItem"
                    title: Options
                    type: array
                text:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Text
                type:
                    anyOf:
                        - $ref: "#/components/schemas/QuestionType"
                        - type: "null"
            required:
                - id
                - assessment_id
                - text
                - type
                - marks
            title: QuestionResponse
            type: object
        QuestionType:
            enum:
                - mcq
                - msq
            title: QuestionType
            type: string
        QuestionUpdate:
            properties:
                marks:
                    anyOf:
                        - type: integer
                        - type: "null"
                    title: Marks
                options:
                    anyOf:
                        - items:
                              $ref: "#/components/schemas/OptionItem"
                          type: array
                        - type: "null"
                    title: Options
                text:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Text
                type:
                    anyOf:
                        - $ref: "#/components/schemas/QuestionType"
                        - type: "null"
            title: QuestionUpdate
            type: object
...
        TestStartResponse:
            description: Returned when a student starts a test (hides correct answers).
            properties:
                assessment_id:
                    title: Assessment Id
                    type: string
                questions:
                    items:
                        $ref: "#/components/schemas/QuestionPublicResponse"
                    title: Questions
                    type: array
                submission_id:
                    title: Submission Id
                    type: string
                title:
                    title: Title
                    type: string
                total_marks:
                    title: Total Marks
                    type: integer
            required:
                - submission_id
                - assessment_id
                - title
                - total_marks
                - questions
            title: TestStartResponse
            type: object
...
        ValidationError:
            properties:
                ctx:
                    title: Context
                    type: object
                input:
                    title: Input
                loc:
                    items:
                        anyOf:
                            - type: string
                            - type: integer
                    title: Location
                    type: array
                msg:
                    title: Message
                    type: string
                type:
                    title: Error Type
                    type: string
            required:
                - loc
                - msg
                - type
            title: ValidationError
            type: object
...
    securitySchemes:
        OAuth2PasswordBearer:
            flows:
                password:
                    scopes: {}
                    tokenUrl: /api/v1/auth/login
            type: oauth2
paths:
...
    /api/v1/notifications/send:
        post:
            operationId: send_notification_api_v1_notifications_send_post
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/NotificationBulkCreate"
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/NotificationResponse"
                                title: Response Send Notification Api V1 Notifications Send Post
                                type: array
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Send a notification to multiple users concurrently (staff only)
            tags:
                - notifications
    /api/v1/notifications/{notification_id}:
        delete:
            operationId: delete_one_api_v1_notifications__notification_id__delete
            parameters:
                - in: path
                  name: notification_id
                  required: true
                  schema:
                      title: Notification Id
                      type: string
            responses:
                "204":
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Delete a notification
            tags:
                - notifications
    /api/v1/notifications/{notification_id}/read:
        patch:
            operationId: mark_read_api_v1_notifications__notification_id__read_patch
            parameters:
                - in: path
                  name: notification_id
                  required: true
                  schema:
                      title: Notification Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/NotificationResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Mark a notification as read
            tags:
                - notifications
    /api/v1/notifications/{user_id}:
        get:
            operationId: list_notifications_api_v1_notifications__user_id__get
            parameters:
                - in: path
                  name: user_id
                  required: true
                  schema:
                      title: User Id
                      type: string
                - description: Number of records to skip
                  in: query
                  name: offset
                  required: false
                  schema:
                      default: 0
                      description: Number of records to skip
                      minimum: 0
                      title: Offset
                      type: integer
                - description: Maximum records to return
                  in: query
                  name: limit
                  required: false
                  schema:
                      default: 50
                      description: Maximum records to return
                      maximum: 1000
                      minimum: 1
                      title: Limit
                      type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Page_NotificationResponse_"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: List all notifications for a user (newest first)
            tags:
                - notifications
...
    /api/v1/questions/{question_id}:
        delete:
            operationId: delete_one_api_v1_questions__question_id__delete
            parameters:
                - in: path
                  name: question_id
                  required: true
                  schema:
                      title: Question Id
                      type: string
            responses:
                "204":
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Delete a question (staff only)
            tags:
                - questions
        patch:
            operationId: update_one_api_v1_questions__question_id__patch
            parameters:
                - in: path
                  name: question_id
                  required: true
                  schema:
                      title: Question Id
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/QuestionUpdate"
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/QuestionResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Update a question (staff only)
            tags:
                - questions
...
    /api/v1/tests/{assessment_id}/start:
        post:
            operationId: start_test_api_v1_tests__assessment_id__start_post
            parameters:
                - in: path
                  name: assessment_id
                  required: true
                  schema:
                      title: Assessment Id
                      type: string
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TestStartResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary:
                "Start a test \u2014 creates a blank submission and returns questions\
                \ (no correct answers)"
            tags:
                - tests
    /api/v1/tests/{assessment_id}/submit:
        post:
            operationId: submit_test_api_v1_tests__assessment_id__submit_post
            parameters:
                - in: path
                  name: assessment_id
                  required: true
                  schema:
                      title: Assessment Id
                      type: string
                - in: query
                  name: submission_id
                  required: true
                  schema:
                      title: Submission Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GradeResult"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary:
                "Submit test answers \u2014 auto-grades MCQ/MSQ, persists score and\
                \ pass/fail"
            tags:
                - tests
...